anvil_zksync_config

Module constants

Constants

DEFAULT_ACCOUNT_BALANCE

Default account balance for the dev accounts

DEFAULT_ESTIMATE_GAS_PRICE_SCALE_FACTOR

Scale factor for estimating L1 gas prices

DEFAULT_ESTIMATE_GAS_SCALE_FACTOR

Scale factor for estimating gas limits

DEFAULT_FAIR_PUBDATA_PRICE

Default price for fair pubdata based on predefined value

DEFAULT_L1_GAS_PRICE

Default L1 gas price for transactions

DEFAULT_L2_GAS_PRICE

Default L2 gas price for transactions if not provided via CLI

DEFAULT_LOG_FILE_PATH

Default log file path for the test node

DEFAULT_MNEMONIC

Default mnemonic phrase for the test node

DERIVATION_PATH

Default derivation path for the test node

EVM_EMULATOR_ENABLER_CALLDATA

Calldata to allow EVM bytecode in ContractDeployer

LEGACY_RICH_WALLETS

List of legacy wallets (address, private key) that we seed with tokens at start.

NODE_PORT

Default port for the test node server

NON_FORK_FIRST_BLOCK_TIMESTAMP

Timestamp of the first block (if not running in fork mode).

PSEUDO_CALLER

Pseudo caller that is allowed to call system-only methods

RICH_WALLETS

TEST_NODE_NETWORK_ID

Network ID for the test node