anvil_zksync_config

Struct TestNodeConfig

pub struct TestNodeConfig {

    pub config_out: Option<String>,
    pub port: u16,
    pub show_node_config: bool,
    pub show_storage_logs: ShowStorageLogs,
    pub show_vm_details: ShowVMDetails,
    pub show_gas_details: ShowGasDetails,
    pub verbosity: u8,
    pub silent: bool,
    pub system_contracts_options: SystemContractsOptions,
    pub system_contracts_path: Option<PathBuf>,
    pub protocol_version: Option<ProtocolVersionId>,
    pub override_bytecodes_dir: Option<String>,
    pub bytecode_compression: bool,
    pub use_evm_interpreter: bool,
    pub zksync_os: ZKsyncOsConfig,
    pub chain_id: Option<u32>,
    pub l1_gas_price: Option<u64>,
    pub l2_gas_price: Option<u64>,
    pub l1_pubdata_price: Option<u64>,
    pub price_scale_factor: Option<f64>,
    pub limit_scale_factor: Option<f32>,
    pub log_level: LogLevel,
    pub log_file_path: String,
    pub cache_dir: String,
    pub cache_config: CacheConfig,
    pub genesis_accounts: Vec<PrivateKeySigner>,
    pub genesis_balance: U256,
    pub account_generator: Option<AccountGenerator>,
    pub signer_accounts: Vec<PrivateKeySigner>,
    pub genesis: Option<Genesis>,
    pub genesis_timestamp: Option<u64>,
    pub enable_auto_impersonate: bool,
    pub offline: bool,
    pub host: Vec<IpAddr>,
    pub health_check_endpoint: bool,
    pub block_time: Option<Duration>,
    pub max_transactions: usize,
    pub no_mining: bool,
    pub allow_origin: String,
    pub no_cors: bool,
    pub transaction_order: TransactionOrder,
    pub state: Option<PathBuf>,
    pub dump_state: Option<PathBuf>,
    pub state_interval: Option<u64>,
    pub preserve_historical_states: bool,
    pub load_state: Option<PathBuf>,
    pub l1_config: Option<L1Config>,
    pub auto_execute_l1: bool,
    pub base_token_config: BaseTokenConfig,

}

Defines the configuration parameters for the [InMemoryNode].

Fields
config_out: Option<String>

Filename to write anvil-zksync output as json

port: u16

Port the node will listen on

show_node_config: bool

Print node config on startup if true

show_storage_logs: ShowStorageLogs

Level of detail for storage logs

show_vm_details: ShowVMDetails

Level of detail for VM execution logs

show_gas_details: ShowGasDetails

Level of detail for gas usage logs

verbosity: u8

Numeric verbosity derived from repeated -v flags (e.g. -v = 1, -vv = 2, etc.).

silent: bool

Don’t print anything on startup if true

system_contracts_options: SystemContractsOptions

Configuration for system contracts

system_contracts_path: Option<PathBuf>

Path to the system contracts directory

protocol_version: Option<ProtocolVersionId>

Protocol version to use for new blocks. Also affects revision of built-in contracts that will get deployed (if applicable)

override_bytecodes_dir: Option<String>

Directory to override bytecodes

bytecode_compression: bool

Enable bytecode compression

use_evm_interpreter: bool

Enables EVM interpreter mode

zksync_os: ZKsyncOsConfig

Enables ZKsyncOS mode (experimental)

chain_id: Option<u32>

Optional chain ID for the node

l1_gas_price: Option<u64>

L1 gas price (optional override)

l2_gas_price: Option<u64>

L2 gas price (optional override)

l1_pubdata_price: Option<u64>

Price for pubdata on L1

price_scale_factor: Option<f64>

L1 gas price scale factor for gas estimation

limit_scale_factor: Option<f32>

The factor by which to scale the gasLimit

log_level: LogLevel

Logging verbosity level

log_file_path: String

Path to the log file

cache_dir: String

Directory to store cache files (defaults to ./cache)

cache_config: CacheConfig

Cache configuration for the test node

genesis_accounts: Vec<PrivateKeySigner>

Signer accounts that will be initialized with genesis_balance in the genesis block.

genesis_balance: U256

Native token balance of every genesis account in the genesis block

account_generator: Option<AccountGenerator>

The generator used to generate the dev accounts

signer_accounts: Vec<PrivateKeySigner>

Signer accounts that can sign messages/transactions

genesis: Option<Genesis>

The genesis to use to initialize the node

genesis_timestamp: Option<u64>

Genesis block timestamp

enable_auto_impersonate: bool

Enable auto impersonation of accounts on startup

offline: bool

Whether the node operates in offline mode

host: Vec<IpAddr>

The host the server will listen on

health_check_endpoint: bool

Whether we need to enable the health check endpoint.

block_time: Option<Duration>

Block time in seconds for interval sealing. If unset, node seals a new block as soon as there is at least one transaction.

max_transactions: usize

Maximum number of transactions per block

no_mining: bool

Disable automatic sealing mode and use BlockSealer::Noop instead

allow_origin: String

The cors allow_origin header

no_cors: bool

Disable CORS if true

transaction_order: TransactionOrder

How transactions are sorted in the mempool

state: Option<PathBuf>

Path to load/dump the state from

dump_state: Option<PathBuf>

Path to dump the state to

state_interval: Option<u64>

Interval to dump the state

preserve_historical_states: bool

Preserve historical states

load_state: Option<PathBuf>

State to load

l1_config: Option<L1Config>

L1 configuration, disabled if None

auto_execute_l1: bool

Whether to automatically execute L1 batches

base_token_config: BaseTokenConfig

Base token configuration

Implementations

impl TestNodeConfig

pub fn protocol_version(&self) -> ProtocolVersionId

impl TestNodeConfig

pub fn print(&self, fork_details: Option<&ForkPrintInfo>)

pub fn set_config_out(self, config_out: Option<String>) -> Self

Sets the file path to write the anvil-zksync config info to.

pub fn with_port(self, port: Option<u16>) -> Self

Set the port for the test node

pub fn get_port(&self) -> u16

Get the port for the test node

pub fn with_chain_id(self, chain_id: Option<u32>) -> Self

Set the chain ID for the test node

pub fn get_chain_id(&self) -> u32

Get the chain ID for the test node

pub fn update_chain_id(&mut self, chain_id: Option<u32>) -> &mut Self

Update the chain ID

pub fn with_system_contracts( self, option: Option<SystemContractsOptions>, ) -> Self

Set the system contracts configuration option

pub fn with_system_contracts_path(self, path: Option<PathBuf>) -> Self

Set the system contracts path

pub fn with_protocol_version( self, protocol_version: Option<ProtocolVersionId>, ) -> Self

Set the protocol version configuration option

pub fn get_system_contracts(&self) -> SystemContractsOptions

Get the system contracts configuration option

pub fn with_override_bytecodes_dir(self, dir: Option<String>) -> Self

Set the override bytecodes directory

pub fn get_override_bytecodes_dir(&self) -> Option<&String>

Get the override bytecodes directory

pub fn with_enforce_bytecode_compression(self, enforce: Option<bool>) -> Self

Set whether bytecode compression is enforced

pub fn is_bytecode_compression_enforced(&self) -> bool

Check if bytecode compression enforcement is enabled

pub fn with_evm_interpreter(self, enable: Option<bool>) -> Self

Enable or disable EVM emulation

pub fn with_zksync_os(self, zksync_os: ZKsyncOsConfig) -> Self

Enable or disable ZKsync OS

pub fn is_evm_interpreter_enabled(&self) -> bool

Get the EVM interpreter status

pub fn with_l1_gas_price(self, price: Option<u64>) -> Self

Set the L1 gas price

pub fn get_l1_gas_price(&self) -> u64

Get the L1 gas price

pub fn update_l1_gas_price(&mut self, price: Option<u64>) -> &mut Self

Update the L1 gas price

pub fn with_l2_gas_price(self, price: Option<u64>) -> Self

Set the L2 gas price

pub fn get_l2_gas_price(&self) -> u64

Get the L2 gas price

pub fn update_l2_gas_price(&mut self, price: Option<u64>) -> &mut Self

Update the L2 gas price

pub fn with_l1_pubdata_price(self, price: Option<u64>) -> Self

Set the L1 pubdata price

pub fn get_l1_pubdata_price(&self) -> u64

Get the L1 pubdata price

pub fn update_l1_pubdata_price(&mut self, price: Option<u64>) -> &mut Self

Update the L1 pubdata price

pub fn with_log_level(self, level: Option<LogLevel>) -> Self

Set the log level

pub fn get_log_level(&self) -> LogLevel

Get the log level

pub fn get_cache_dir(&self) -> &str

Gets the cache directory

pub fn with_cache_dir(self, dir: Option<String>) -> Self

Set the cache directory

pub fn with_cache_config(self, config: Option<CacheConfig>) -> Self

Set the cache configuration

pub fn get_cache_config(&self) -> &CacheConfig

Get the cache configuration

pub fn with_log_file_path(self, path: Option<String>) -> Self

Set the log file path

pub fn get_log_file_path(&self) -> &str

Get the log file path

pub fn with_verbosity_level(self, verbosity: u8) -> Self

Sets the numeric verbosity derived from repeated -v flags

pub fn get_verbosity_level(&self) -> u8

Get the numeric verbosity derived from repeated -v flags

pub fn with_silent(self, silent: Option<bool>) -> Self

Enable or disable silent mode

pub fn with_show_node_config(self, show_node_config: Option<bool>) -> Self

Enable or disable printing node config on startup

pub fn with_show_storage_logs( self, show_storage_logs: Option<ShowStorageLogs>, ) -> Self

Set the visibility of storage logs

pub fn get_show_storage_logs(&self) -> ShowStorageLogs

Get the visibility of storage logs

pub fn with_vm_log_detail(self, detail: Option<ShowVMDetails>) -> Self

Set the detail level of VM execution logs

pub fn get_vm_log_detail(&self) -> ShowVMDetails

Get the detail level of VM execution logs

pub fn with_show_gas_details( self, show_gas_details: Option<ShowGasDetails>, ) -> Self

Set the visibility of gas usage logs

pub fn get_show_gas_details(&self) -> ShowGasDetails

Get the visibility of gas usage logs

pub fn with_gas_limit_scale(self, scale: Option<f32>) -> Self

Set the gas limit scale factor

pub fn get_gas_limit_scale(&self) -> f32

Get the gas limit scale factor

pub fn update_gas_limit_scale(&mut self, scale: Option<f32>) -> &mut Self

Update the gas limit scale factor

pub fn with_price_scale(self, scale: Option<f64>) -> Self

Set the price scale factor

pub fn get_price_scale(&self) -> f64

Get the price scale factor

pub fn update_price_scale(&mut self, scale: Option<f64>) -> &mut Self

Updates the price scale factor

pub fn with_genesis_balance<U: Into<U256>>(self, balance: U) -> Self

Sets the balance of the genesis accounts in the genesis block

pub fn with_genesis_accounts(self, accounts: Vec<PrivateKeySigner>) -> Self

Sets the genesis accounts.

pub fn with_signer_accounts(self, accounts: Vec<PrivateKeySigner>) -> Self

Sets the signer accounts

pub fn with_account_generator(self, generator: AccountGenerator) -> Self

Sets both the genesis accounts and the signer accounts so that genesis_accounts == accounts

pub fn with_genesis_timestamp(self, timestamp: Option<u64>) -> Self

Sets the genesis timestamp

pub fn get_genesis_timestamp(&self) -> u64

Returns the genesis timestamp to use

pub fn with_genesis(self, genesis: Option<Genesis>) -> Self

Sets the init genesis (genesis.json)

pub fn with_auto_impersonate(self, enable_auto_impersonate: bool) -> Self

Sets whether to enable autoImpersonate

pub fn with_offline(self, offline: Option<bool>) -> Self

Set the offline mode

pub fn is_offline(&self) -> bool

Get the offline mode status

pub fn with_host(self, host: Vec<IpAddr>) -> Self

Sets the host the server will listen on

pub fn with_health_check_endpoint( self, health_check_endpoint: Option<bool>, ) -> Self

Set the health check endpoint mode

pub fn is_health_check_endpoint_endpoint_enabled(&self) -> bool

Get the health check endpoint mode status

pub fn with_block_time(self, block_time: Option<Duration>) -> Self

Set the block time

pub fn with_no_mining(self, no_mining: bool) -> Self

If set to true auto sealing will be disabled

pub fn with_transaction_order(self, transaction_order: TransactionOrder) -> Self

pub fn with_allow_origin(self, allow_origin: String) -> Self

Set allow_origin CORS header

pub fn with_no_cors(self, no_cors: bool) -> Self

Enable or disable CORS

pub fn with_state(self, state: Option<PathBuf>) -> Self

Set the state

pub fn with_dump_state(self, dump_state: Option<PathBuf>) -> Self

Set the state dump path

pub fn with_state_interval(self, state_interval: Option<u64>) -> Self

Set the state dump interval

pub fn with_preserve_historical_states( self, preserve_historical_states: bool, ) -> Self

Set preserve historical states

pub fn with_load_state(self, load_state: Option<PathBuf>) -> Self

Set the state to load

pub fn with_l1_config(self, l1_config: Option<L1Config>) -> Self

Set the L1 config

pub fn with_auto_execute_l1(self, auto_execute_l1: Option<bool>) -> Self

Set the auto L1 execution

pub fn with_base_token_config(self, base_token_config: BaseTokenConfig) -> Self

Set the base token config